requirements.txt
requests
beautifulsoup4
camelot-py[cv]
pandas
